+++ yasio/config.hpp
@@ -205,7 +205,7 @@ SOFTWARE.
 /*
 **  The yasio version macros
 */
-#define YASIO_VERSION_NUM 0x040105
+#define YASIO_VERSION_NUM 0x050001
 
 /*
 ** The macros used by io_service.