+++ yasio/impl/openssl.hpp
@@ -31,6 +31,8 @@ SOFTWARE.
 
 #if YASIO_SSL_BACKEND == 1 // OpenSSL
 
+#  include "yasio/split.hpp"
+
 // The ssl error mask (1 << 31), a little hack, but works
 #  define YSSL_ERR_MASK 0x80000000
 
@@ -51,8 +53,8 @@ YASIO__DECL yssl_ctx_st* yssl_ctx_new(const yssl_options& opts)
     if (yasio__valid_str(opts.crtfile_))
     { // CAfile for verify
       fail_count = 0;
-      yssl_splitpath(opts.crtfile_, [&](char* first, char* last) {
-        yssl_split_term null_term(last);
+      yasio::split(opts.crtfile_, ',', [&](char* first, char* last) {
+        yasio::split_term null_term(last);
 
 #  if defined(OPENSSL_VERSION_MAJOR) && (OPENSSL_VERSION_MAJOR >= 3)
         /* OpenSSL 3.0.0 has deprecated SSL_CTX_load_verify_locations */
@@ -65,15 +67,13 @@ YASIO__DECL yssl_ctx_st* yssl_ctx_new(const yssl_options& opts)
           ++fail_count;
           YASIO_LOG("[global] load ca certifaction file failed!");
         }
-
-        return !ok;
       });
     }
     /*
-    * client cert & key not implement yet, since it not common usecase
-    * SSL_CTX_use_certificate_chain_file
-    * SSL_CTX_use_PrivateKey_file
-    */
+     * client cert & key not implement yet, since it not common usecase
+     * SSL_CTX_use_certificate_chain_file
+     * SSL_CTX_use_PrivateKey_file
+     */
     if (!fail_count)
     {
       ::SSL_CTX_set_verify(ctx, SSL_VERIFY_PEER, ::SSL_CTX_get_verify_callback(ctx));
@@ -99,12 +99,12 @@ YASIO__DECL yssl_ctx_st* yssl_ctx_new(const yssl_options& opts)
     if (yasio__valid_str(opts.keyfile_) && ::SSL_CTX_use_PrivateKey_file(ctx, opts.keyfile_, SSL_FILETYPE_PEM) <= 0)
       YASIO_LOG("[gobal] load server private key file failed!");
 
-    /* 
-    * If client provide cert, then verify, otherwise skip verify
-    * Note: if SSL_VERIFY_FAIL_IF_NO_PEER_CERT specified, client must provide
-    * cert & key which is not common use, means 100% online servers doesn't require
-    * client to provide cert
-    */
+    /*
+     * If client provide cert, then verify, otherwise skip verify
+     * Note: if SSL_VERIFY_FAIL_IF_NO_PEER_CERT specified, client must provide
+     * cert & key which is not common use, means 100% online servers doesn't require
+     * client to provide cert
+     */
     ::SSL_CTX_set_verify(ctx, SSL_VERIFY_PEER, ::SSL_CTX_get_verify_callback(ctx));
   }